requirements.txt
PyYAML>=6.0
ansible-core>=2.14.0
